--- Assets/Project/Scenes/SceneTest/DaichiTest/PlayerHealth.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 9a6cb425c2947db49abccff9fc2bb14b
