+++ Assets/Project/Scenes/SceneTest/SakuraScene.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: d124ae1618a70f543b1bfb2bd45fb2d1
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
