+++ Assets/Project/Scenes/SceneTest/SakuraTest.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 32fecbde026de304993ff6848405984b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
